requirements.txt
torch
torchvision
pillow
pandas
numpy
matplotlib
scikit-learn
streamlit
